move sharedpreferencehelper to preference package
