+++ app/src/main/java/org/stepik/android/adaptive/api/Api.kt
@@ -4,7 +4,7 @@ import android.net.Uri
 import org.stepik.android.adaptive.configuration.Config
 import org.stepik.android.adaptive.Util
 import org.stepik.android.adaptive.api.auth.SocialManager
-import org.stepik.android.adaptive.data.SharedPreferenceHelper
+import org.stepik.android.adaptive.data.preference.SharedPreferenceHelper
 import org.stepik.android.adaptive.data.model.EnrollmentWrapper
 import org.stepik.android.adaptive.data.model.AccountCredentials
 import org.stepik.android.adaptive.data.model.Profile
