+++ app/src/main/java/org/stepik/android/adaptive/api/storage/RemoteStorageRepositoryImpl.kt
@@ -8,7 +8,7 @@ import org.stepik.android.adaptive.api.StepikService
 import org.stepik.android.adaptive.api.storage.model.StorageRequest
 import org.stepik.android.adaptive.api.storage.model.StorageResponse
 import org.stepik.android.adaptive.configuration.Config
-import org.stepik.android.adaptive.data.SharedPreferenceHelper
+import org.stepik.android.adaptive.data.preference.SharedPreferenceHelper
 import org.stepik.android.adaptive.data.model.QuestionsPackStorageItem
 import org.stepik.android.adaptive.data.model.StorageRecord
 import org.stepik.android.adaptive.di.AppSingleton
