+++ app/src/main/java/org/stepik/android/adaptive/content/questions/QuestionsPacksManager.kt
@@ -1,7 +1,7 @@
 package org.stepik.android.adaptive.content.questions
 
 import org.stepik.android.adaptive.data.Analytics
-import org.stepik.android.adaptive.data.SharedPreferenceHelper
+import org.stepik.android.adaptive.data.preference.SharedPreferenceHelper
 import org.stepik.android.adaptive.di.AppSingleton
 import javax.inject.Inject
