+++ app/src/main/java/org/stepik/android/adaptive/data/preference/SharedPreferencesExtensions.kt
@@ -1,4 +1,4 @@
-package org.stepik.android.adaptive.data
+package org.stepik.android.adaptive.data.preference
 
 import android.content.SharedPreferences
 import kotlin.properties.ReadOnlyProperty
