+++ app/src/main/java/org/stepik/android/adaptive/gamification/ExpManager.kt
@@ -3,7 +3,7 @@ package org.stepik.android.adaptive.gamification
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig
 import org.stepik.android.adaptive.api.Api
 import org.stepik.android.adaptive.data.Analytics
-import org.stepik.android.adaptive.data.SharedPreferenceHelper
+import org.stepik.android.adaptive.data.preference.SharedPreferenceHelper
 import org.stepik.android.adaptive.data.db.DataBaseMgr
 
 import io.reactivex.Completable
