+++ app/src/main/java/org/stepik/android/adaptive/ui/activity/SplashActivity.kt
@@ -11,7 +11,7 @@ import io.reactivex.functions.BiFunction
 import org.stepik.android.adaptive.App
 import org.stepik.android.adaptive.R
 import org.stepik.android.adaptive.core.ScreenManager
-import org.stepik.android.adaptive.data.SharedPreferenceHelper
+import org.stepik.android.adaptive.data.preference.SharedPreferenceHelper
 import org.stepik.android.adaptive.di.qualifiers.BackgroundScheduler
 import org.stepik.android.adaptive.di.qualifiers.MainScheduler
 import java.util.concurrent.TimeUnit
