+++ app/src/main/java/org/stepik/android/adaptive/util/RatingNamesGenerator.kt
@@ -2,7 +2,7 @@ package org.stepik.android.adaptive.util
 
 import android.content.Context
 import org.stepik.android.adaptive.R
-import org.stepik.android.adaptive.data.SharedPreferenceHelper
+import org.stepik.android.adaptive.data.preference.SharedPreferenceHelper
 import org.stepik.android.adaptive.di.AppSingleton
 import javax.inject.Inject
